replaced doubles with floats

+++ src/tab.hpp
@@ -10,7 +10,7 @@ public:
     virtual void draw(int w, int h) = 0;
     virtual void handleInput(unsigned int input) = 0;
 
-    void volumeBar(int w, int h, int px, int py, double vol, double peak);
+    void volumeBar(int w, int h, int px, int py, float vol, float peak);
 private:
     void fillW(int w, int h, int offset_x, int offset_y, const char *str);
     unsigned int getVolumeColor(int p);
